Automate Creation/Destruction

+++ backend_code/DAO.py
@@ -60,6 +60,12 @@ class Account(DAO):
         if not self._write_access: return False
         return self._password
 
+    def get_cart( self ):
+        if not self._read_access: return False
+        cart = ShoppingCart(self)
+        cart.load()
+        return cart
+
     def set_password( self, password ):
         if self._write_access:
             self._password = password
@@ -68,44 +74,46 @@ class Account(DAO):
 
     def create( self ):
         if self._password:
-            sql = "INSERT INTO Account VALUES (%s,%s)"
-            create_tuple = (self._username,self._password)
-            if db_accessor.run_change( sql, create_tuple ):
+            if db_accessor.run_change(
+                    "INSERT INTO Account VALUES (%s,%s)",
+                    self._username, self._password ):
                 self._access_set_all()
+                ShoppingCart(self).create()
                 return True
         return False
 
     def update( self ):
         if not self._write_access: return False
-        sql = "UPDATE Account SET password=%s WHERE username=%s"
-        update_tuple = (self._password,self._username)
-        return db_accessor.run_change(sql,update_tuple)
+        return db_accessor.run_change(
+                "UPDATE Account SET password=%s WHERE username=%s",
+                self._password,self._username)
 
     def remove( self ):
         if not self._write_access: return False
-        sql = "DELETE FROM Account WHERE username=%s"
-        remove_tuple = (self._username,)
-        if db_accessor.run_change( sql, remove_tuple ):
-            self._access_remove()
+        if self.get_cart().remove():
+            db_accessor.run_change(
+                    "DELETE FROM Account WHERE username=%s",
+                    self._username )
             return True
         return False
 
     def load( self ):
         # intent = read and write
         if self._password:
-            sql = "SELECT * FROM Account WHERE username=%s AND password=%s"
-            load_tuple = (self._username,self._password)
-            if db_accessor.run_select(sql,load_tuple):
+            if db_accessor.run_select(
+                    "SELECT * FROM Account WHERE username=%s AND password=%s",
+                    self._username,self._password):
                 self._access_set_all()
                 return True
         # intent = write only
         else:
-            sql = "SELECT * FROM Account WHERE username=%s"
-            load_tuple = (self._username,)
-            if db_accessor.run_select(sql,load_tuple):
+            if db_accessor.run_select(
+                    "SELECT * FROM Account WHERE username=%s",
+                    self._username):
                 self._access_set_readonly()
                 return True
         return False
+
 
 class ShoppingCart(DAO):
     def __init__( self, account ):
@@ -117,9 +125,9 @@ class ShoppingCart(DAO):
         return self._account.get_username()
 
     def create( self ):
-        sql = "INSERT INTO ShoppingCart VALUES (%s)"
-        create_tuple = (self._account.get_username(),)
-        if db_accessor.run_change(sql,create_tuple):
+        if db_accessor.run_change(
+                "INSERT INTO ShoppingCart VALUES (%s)",
+                    self._account.get_username()):
             self._access_set_by_reference( self._account )
             return True
         return False
@@ -129,9 +137,10 @@ class ShoppingCart(DAO):
 
     def remove( self ):
         if self._write_access:
-            sql = "DELETE FROM ShoppingCart WHERE id=%s"
-            remove_tuple = (self.get_id(),)
-            if db_accessor.run_change(sql,remove_tuple):
+            self.clear()
+            if db_accessor.run_change(
+                    "DELETE FROM ShoppingCart WHERE id=%s",
+                    self.get_id()):
                 self._access_remove()
                 return True
         return False
@@ -140,6 +149,26 @@ class ShoppingCart(DAO):
         if not self._account.write_access(): return False
         self._access_set_by_reference( self._account )
         return True
+
+    def get_item_selections( self ):
+        if not self.read_access(): return False
+        selections = list()
+        selection_results = db_accessor.run_select(
+                                "SELECT * FROM ItemSelection WHERE cart_id=%s",
+                                self.get_id())
+        if selection_results:
+            for item_id,cart_id,quantity in selection_results:
+                item = Item(item_id)
+                item.load()
+                selection = ItemSelection(self,item)
+                selection.load()
+                selections.append( selection )
+        return selections
+
+    def clear( self ):
+        if not self.write_access(): return False
+        for selection in self.get_item_selections():
+            selection.remove()
 
 
 class Item(DAO):
@@ -170,17 +199,17 @@ class Item(DAO):
         raise NotImplemented("Currently write access does not exist for Item\nInstead create a new Item")
 
     def create( self ):
-        sql = "INSERT INTO Item VALUES (%s,%s,%s)"
-        create_tuple = (self._id,self._name,self._source)
-        if db_accessor.run_change(sql,create_tuple):
+        if db_accessor.run_change(
+                    "INSERT INTO Item VALUES (%s,%s,%s)",
+                    self._id,self._name,self._source):
             self._access_set_readonly()
             return True
         return False
 
     def load( self ):
-        sql = "SELECT * FROM Item WHERE id=%s"
-        load_tuple = (self._id,)
-        selection_result = db_accessor.run_select(sql,load_tuple)
+        selection_result = db_accessor.run_select(
+                                "SELECT * FROM Item WHERE id=%s",
+                                self._id)
         if selection_result:
             item_id, item_name, item_source = selection_result[0]
             self._id = item_id
@@ -222,17 +251,17 @@ class ItemSelection( DAO ):
         self._quantity = quantity
 
     def create( self ):
-        sql = "INSERT INTO ItemSelection VALUES (%s,%s,%s)"
-        create_tuple = (self._item.get_id(),self._cart.get_id(),self._quantity)
-        if db_accessor.run_change(sql,create_tuple):
+        if db_accessor.run_change(
+                    "INSERT INTO ItemSelection VALUES (%s,%s,%s)",
+                    self._item.get_id(),self._cart.get_id(),self._quantity):
             self._access_set_by_reference( self._cart )
             return True
         return False
 
     def load( self ):
-        sql = "SELECT * FROM ItemSelection WHERE item_id=%s AND cart_id=%s"
-        load_tuple = (self._item.get_id(),self._cart.get_id())
-        selection_result = db_accessor.run_select(sql,load_tuple)
+        selection_result = db_accessor.run_select(
+                    "SELECT * FROM ItemSelection WHERE item_id=%s AND cart_id=%s",
+                    self._item.get_id(),self._cart.get_id())
         if selection_result:
             item_id,cart_id,quantity = selection_result[0]
             self._quantity = quantity
@@ -242,16 +271,17 @@ class ItemSelection( DAO ):
 
     def update( self ):
         if not self._write_access: return False
-        sql = "UPDATE ItemSelection SET quantity=%s WHERE cart_id=%s AND item_id=%s"
-        update_tuple = (self.get_quantity(),self.get_cart().get_id(),
-                        self.get_item().get_id())
-        return db_accessor.run_change(sql,update_tuple)
+        return db_accessor.run_change(
+                    "UPDATE ItemSelection SET quantity=%s WHERE cart_id=%s AND item_id=%s",
+                    self.get_quantity(),
+                    self.get_cart().get_id(),
+                    self.get_item().get_id())
 
     def remove( self ):
         if not self._write_access: return False
-        sql = "DELETE FROM ItemSelection WHERE cart_id=%s AND item_id=%s"
-        remove_tuple = (self.get_cart().get_id(),self.get_item().get_id())
-        if db_accessor.run_change( sql, remove_tuple ):
+        if db_accessor.run_change(
+                    "DELETE FROM ItemSelection WHERE cart_id=%s AND item_id=%s",
+                    self.get_cart().get_id(),self.get_item().get_id()):
             self._access_remove()
             return True
         return False
@@ -292,13 +322,12 @@ if __name__ == "__main__":
     test_selections = True
     if test_selections:
         print( "Create selection" )
-        # create Michael's Pizza
+        # create Michael's item
         my_account = Account("Michael", "my_password")
         assert my_account.create() == True
 
-        # TODO: automatically generate cart on account creation
         my_cart = ShoppingCart( my_account )
-        assert my_cart.create() == True
+        assert my_cart.load() == True
 
         my_item = Item(42,"My Item","https://my_source.com")
         if not my_item.load():
@@ -309,7 +338,7 @@ if __name__ == "__main__":
         assert my_selection.create() == True
 
         print( "Access selection" )
-        # access Michael's Pizza
+        # access Michael's item
         my_account = Account("Michael", "my_password")
         assert my_account.load() == True
         my_cart = ShoppingCart( my_account )
@@ -321,15 +350,11 @@ if __name__ == "__main__":
         my_selection = ItemSelection( my_cart, my_item )
         assert my_selection.load() == True
         assert my_selection.get_quantity() == 5 # verify correct amount
-        assert my_selection.remove() == True
 
         # delete Michael's account
-        # ERROR: must delete ShoppingCart first
-        #assert my_account.remove() == True
-        db_accessor.run_change( "DELETE FROM ItemSelection WHERE cart_id=%s", ("Michael",) )
-        db_accessor.run_change( "DELETE FROM ShoppingCart WHERE id=%s", ("Michael",) )
-        db_accessor.run_change( "DELETE FROM Account WHERE username=%s", ("Michael",) )
-        db_accessor.run_change( "DELETE FROM Item WHERE id=%s", ("42",) )
+        assert my_account.remove() == True
+
+
